flask/requirements.txt
alembic==1.0.10
asn1crypto==0.24.0
blinker==1.4
certifi==2019.6.16
cffi==1.12.3
Click==7.0
cryptography==2.7
dominate==2.3.5
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==0.6.1
Mako==1.0.12
MarkupSafe==1.1.1
numpy==1.16.4
psycopg2==2.7.6.1
pycparser==2.19
python-dateutil==2.8.0
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.5
visitor==0.1.3
Werkzeug==0.15.4
WTForms==2.2.1
uWSGI==2.0.18
python-dotenv==0.10.3